flatten the spec more
