+++ Makefile
@@ -30,19 +30,19 @@ openapi:
 	-e GOPATH=/go \
 	-v $(CURDIR):/go/src/github.com/mailchain/mailchain \
 	-w /go/src/github.com/mailchain/mailchain \
-	mailchain/goswagger-tool swagger generate spec -b ./cmd/mailchain/internal/rest -o ./docs/openapi/spec.json
+	mailchain/goswagger-tool swagger generate spec -b ./cmd/mailchain/internal/http -o ./docs/openapi/spec.json
 
 	echo "" >>  ./docs/openapi/spec.json
 
-	echo "package spec" >  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo "" >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo "// nolint: lll" >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo 'func spec() string {' >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo '  return `' >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	cat  ./docs/openapi/spec.json >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo '`' >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	echo '}' >>  ./cmd/mailchain/internal/rest/spec/openapi.go
-	addlicense -l apache -c Finobo ./cmd/mailchain/internal/rest/spec/openapi.go	
+	echo "package handlers" >  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo "" >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo "// nolint: lll" >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo 'func spec() string {' >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo '  return `' >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	cat  ./docs/openapi/spec.json >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo '`' >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	echo '}' >>  ./cmd/mailchain/internal/http/spec/openapi.go
+	addlicense -l apache -c Finobo ./cmd/mailchain/internal/http/spec/openapi.go	
 	rm -rf vendor
 	
 snapshot:
