+++ cmd/mailchain/internal/http/errs/writer.go
