+++ cmd/mailchain/internal/http/errs/writer_test.go
